+++ src/main.js
@@ -1,4 +1,4 @@
-import { VideoBackgrounds, VideoBackgroundGroup, SeekBar } from './video-backgrounds.js';
+import { VideoBackgrounds } from './video-backgrounds.js';
 
 if (typeof jQuery == 'function') {
   (function ($) {
@@ -15,5 +15,3 @@ if (typeof jQuery == 'function') {
 }
 
 window.VideoBackgrounds = VideoBackgrounds;
-window.VideoBackgroundGroup = VideoBackgroundGroup;
-window.SeekBar = SeekBar;
